requirements.txt
google-cloud-asset
google-api-python-client
graphviz